create helper for frienships, move sessions helper from app controller into its own helper file, implment friend? method checks to hide the add friend button when appropriate and avoid creation of friendships that may already exist.
