+++ app/controllers/application_controller.rb
@@ -1,13 +1,7 @@
 class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
 
-  def current_user
-    @current_user ||= User.find(session[:user_id]) if session[:user_id]
-  end
-  helper_method :current_user
-
-  def authorize
-    redirect_to '/' unless current_user
-  end
+  include FriendshipsHelper
+  include SessionsHelper
 
 end
