+++ app/helpers/sessions_helper.rb
@@ -1,0 +1,9 @@
+module SessionsHelper
+  def current_user
+    @current_user ||= User.find(session[:user_id]) if session[:user_id]
+  end
+
+  def authorize
+    redirect_to '/' unless current_user
+  end
+end
